# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Lore Stellar blockchain token explorer. Provides access to MTLAP (Persons), MTLAC (Companies), and MTLAX (Synthetic) accounts, relationships, reputation scores, and statistics.",
        "title": "Lore API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "description": "Returns a paginated list of accounts, optionally filtered by type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List accounts",
                "parameters": [
                    {
                        "enum": [
                            "person",
                            "corporate",
                            "synthetic"
                        ],
                        "type": "string",
                        "description": "Account type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "description": "Returns full account detail including metadata, trustlines, LP shares, trust ratings, and reputation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stellar account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.AccountDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{id}/relationships": {
            "get": {
                "description": "Returns relationships grouped by category (family, work, network, ownership, social)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "relationships"
                ],
                "summary": "Get account relationships",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stellar account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by relationship type (e.g. Spouse, Employer)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter to only confirmed relationships",
                        "name": "confirmed",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter to only mutual relationships",
                        "name": "mutual",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.RelationshipCategoryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{id}/reputation": {
            "get": {
                "description": "Returns the reputation graph including direct raters and raters of raters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reputation"
                ],
                "summary": "Get reputation graph",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stellar account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReputationGraphResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/search": {
            "get": {
                "description": "Search accounts by name or account ID, optionally filtered by tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (min 2 chars)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tag names",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "balance",
                            "reputation"
                        ],
                        "type": "string",
                        "default": "balance",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stats": {
            "get": {
                "description": "Returns aggregate statistics for accounts, persons, companies, and synthetic tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get aggregate statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AccountDetailResponse": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.RelationshipCategoryResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_corporate": {
                    "type": "boolean"
                },
                "lp_shares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LPShareResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "reputation": {
                    "$ref": "#/definitions/api.ReputationResponse"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_xlm_value": {
                    "type": "number"
                },
                "trust_rating": {
                    "$ref": "#/definitions/api.TrustRatingResponse"
                },
                "trustlines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TrustlineResponse"
                    }
                },
                "websites": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.LPShareResponse": {
            "type": "object",
            "properties": {
                "pool_id": {
                    "type": "string"
                },
                "reserve_a": {
                    "$ref": "#/definitions/api.ReserveResponse"
                },
                "reserve_b": {
                    "$ref": "#/definitions/api.ReserveResponse"
                },
                "share_balance": {
                    "type": "string"
                },
                "share_percent": {
                    "type": "string"
                },
                "xlm_value": {
                    "type": "number"
                }
            }
        },
        "api.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {
                    "$ref": "#/definitions/api.Pagination"
                }
            }
        },
        "api.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.RelationshipCategoryResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "relationships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.RelationshipResponse"
                    }
                }
            }
        },
        "api.RelationshipResponse": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "string"
                },
                "is_confirmed": {
                    "type": "boolean"
                },
                "is_mutual": {
                    "type": "boolean"
                },
                "target_id": {
                    "type": "string"
                },
                "target_name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.ReputationGraphResponse": {
            "type": "object",
            "properties": {
                "level1_nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ReputationNodeResponse"
                    }
                },
                "level2_nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ReputationNodeResponse"
                    }
                },
                "score": {
                    "$ref": "#/definitions/api.ReputationResponse"
                },
                "target_account_id": {
                    "type": "string"
                },
                "target_name": {
                    "type": "string"
                }
            }
        },
        "api.ReputationNodeResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "connections": {
                    "type": "integer"
                },
                "distance": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "own_score": {
                    "type": "number"
                },
                "portfolio_xlm": {
                    "type": "number"
                },
                "rating": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "api.ReputationResponse": {
            "type": "object",
            "properties": {
                "base_score": {
                    "type": "number"
                },
                "grade": {
                    "type": "string"
                },
                "rating_count_a": {
                    "type": "integer"
                },
                "rating_count_b": {
                    "type": "integer"
                },
                "rating_count_c": {
                    "type": "integer"
                },
                "rating_count_d": {
                    "type": "integer"
                },
                "total_ratings": {
                    "type": "integer"
                },
                "total_weight": {
                    "type": "number"
                },
                "weighted_score": {
                    "type": "number"
                }
            }
        },
        "api.ReserveResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "asset_code": {
                    "type": "string"
                },
                "asset_issuer": {
                    "type": "string"
                }
            }
        },
        "api.StatsResponse": {
            "type": "object",
            "properties": {
                "total_accounts": {
                    "type": "integer"
                },
                "total_companies": {
                    "type": "integer"
                },
                "total_persons": {
                    "type": "integer"
                },
                "total_synthetic": {
                    "type": "integer"
                },
                "total_xlm_value": {
                    "type": "number"
                }
            }
        },
        "api.TrustRatingResponse": {
            "type": "object",
            "properties": {
                "count_a": {
                    "type": "integer"
                },
                "count_b": {
                    "type": "integer"
                },
                "count_c": {
                    "type": "integer"
                },
                "count_d": {
                    "type": "integer"
                },
                "grade": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.TrustlineResponse": {
            "type": "object",
            "properties": {
                "asset_code": {
                    "type": "string"
                },
                "asset_issuer": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "limit": {
                    "type": "string"
                }
            }
        }
    }
}